Add components for each destination (#5)
This adds a basic (Claude-styled) component `DestinationCard`. This is
not the final design, and it still is missing the `goalStatus` and color
coding. But this gets the overall frontend code setup so I want to land
it as a checkpoint.

Note that I intentionally don't DRY the `routes` because there is enough
divergence that I think having a helper function would make things more
confusing, like more indirection.