Add UI for different travel time states (#13)
We use universal reactivity
(https://svelte.dev/tutorial/svelte/universal-reactivity) to have global
state for the status of travel times. In a follow-up PR, the Address
component will update the value.

This PR is focused on implementing the four distinct UI states (thanks
Claude). We model the four states with an algebraic data type.