Set consistent departure time for transit (#19)
Now,c allers set the day of the week and time, like Monday at 9 am. 

Google Maps Route API lets us specify up to 100 days in the future, so
we always use a future date within the current or following week.

Using Luxon is important so that we can consistently use ET as the time
zone, rather than using UTC or local time. This is feasible because the
app is hardcoded to NYC. It's important to not use local time because I
don't know where the server will be deployed.